# api/openapi.yaml
openapi: 3.0.3
info:
  title: IMS ERP System API
  description: |
    Complete ERP system API with microservices architecture
    
    ## Architecture
    - **API Gateway** (8080) - Central entry point for all client requests
    - **Auth Service** (8081) - Authentication and authorization
    - **Client Services** (8082) - Client management (CQRS pattern)
    - **Invoice Service** (8083) - Invoice management
    - **Payment Service** (8084) - Payment processing
    - **Product Service** (8085) - Product catalog
    - **Order Service** (8086) - Order management
    - **Inventory Service** (8087) - Inventory control
    
    ## Authentication
    All protected endpoints require Bearer token authentication.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited to 1000 requests per minute per tenant.
    
    ## Error Handling
    All errors follow standard HTTP status codes with detailed error messages.
  version: 1.0.0
  contact:
    name: IMS ERP Support
    email: support@ims-erp.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.ims-erp.example.com/api/v1
    description: Production server
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account in the specified tenant
      parameters:
        - in: query
          name: tenantId
          required: true
          schema:
            type: string
            format: uuid
          description: Tenant ID for the new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates user and returns access/refresh tokens
      parameters:
        - in: query
          name: tenantId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
                - refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
...
components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/TokenPair'
        sessionId:
          type: string
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          default: Bearer
        expiresIn:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [active, inactive, locked]
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                default: UNAUTHORIZED
              message:
                type: string
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                default: ALREADY_EXISTS
              message:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []